updates of 1st July

+++ evaluate_ate.py
@@ -6,7 +6,7 @@ import pandas as pd
 import numpy as np
 from tqdm import tqdm
 import matplotlib.pyplot as plt
-from eval_utils import get_kde, get_kde_log_likelihood,plot_kde,calculate_wasserstein_distance,get_distribution_metrics
+from eval_utils import get_kde, get_kde_log_likelihood,plot_kde,calculate_wasserstein_distance,get_distribution_metrics,get_ate_precision_recall
 
 def read_pickle_path(path_):
     with open(path_,'rb') as f:
@@ -33,7 +33,8 @@ if __name__=="__main__":
     seed_number = int(sys_args[1])
     #if seed_number != 0:
     #SEED_TO_USE = [i for i in range(seed_number-5,seed_number)]
-    SEED_TO_USE = [i for i in range(26)]
+    #SEED_TO_USE = [i for i in range(26)]
+    SEED_TO_USE = [seed_number]
 
     BASELINE_FOLDER = str(sys_args[4])
     ATE_FOLDER = str(sys_args[5])
@@ -48,6 +49,7 @@ if __name__=="__main__":
     ate_dataframes = [pd.read_csv(f) for f in ate_csvs]
     ate_dataframe_concatenated = pd.concat(ate_dataframes) # this is the one main dataframe.
     start_time = time.time()
+    REDO = False # whether to redo the KDE CSV file even if it exists.
     with tqdm(total = len(baseline_to_use)*len(SEED_TO_USE)) as pbar:
         for baseline in baseline_to_use:
             for seed in SEED_TO_USE:
@@ -66,7 +68,7 @@ if __name__=="__main__":
                             estimate_ate_samples = ate_of_interest['ates'].values.tolist()
                             # remove nan values in the list
                             estimate_ate_samples = [s for s in estimate_ate_samples if np.isnan(s) == False and np.isinf(s) == False]
-
+                            #breakpoint()
                             if estimate_ate_samples == []:
                                 print(f'Estimate ATE samples are empty. Moving on... \n sys args | baseline: {baseline}, seed:{seed}, treatment:{treatment}, outcome: {outcome}')
                                 continue
@@ -87,6 +89,16 @@ if __name__=="__main__":
                                 print(f'True ATE samples are empty. Moving on... \n sys args: {sys_args}')
                                 continue
 
+
+                            # if we are in bcdnets, and sample size is not 1000, then randomly sample 1000 with replacement
+                            if baseline=='bcdnets':
+                                if len(estimate_ate_samples)<1000:
+                                    estimate_ate_samples = np.random.choice(estimate_ate_samples,1000,replace=True).tolist()
+                                    REDO = True
+                                else:
+                                    REDO = False
+
+
                             precision,recall,metrics = get_distribution_metrics(estimate_ate_samples,true_ate_samples)
                             precision_0,recall_0,metrics_0 = get_ate_precision_recall(estimate_ate_samples,true_ate_samples,0.0)
                             precision_025,recall_025,metrics_025 = get_ate_precision_recall(estimate_ate_samples,true_ate_samples,0.025)
@@ -100,23 +112,42 @@ if __name__=="__main__":
                             KDE_MODEL_FILENAME = os.path.join(SCRATCH_FOLDER,f'{baseline}_{seed}_{treatment}_{outcome}_kde.pkl')
 
 
-                            if os.path.exists(KDE_EVALUATION_DETAILS_FILENAME):
+                            if os.path.exists(KDE_EVALUATION_DETAILS_FILENAME) and REDO == False:
+                                REDO = False
                                 # Check that the KDE is not using bandwidth 0.001
                                 kde_old = read_pickle_path(KDE_MODEL_FILENAME)
                                 if kde_old.bandwidth==0.001:
                                     print(f'KDE exists and has bandwidth of 0.001..')
                                     kde_eval_df = pd.read_csv(KDE_EVALUATION_DETAILS_FILENAME)
-                                    if 'precision' in kde_eval_df:
+                                    if 'precision_0' in kde_eval_df:
                                         print('Already found precision and recall metrics. Skipping...')
                                         continue
                                     else:
-                                        print(f'Adding precision and recall metrics...')
+                                        print(f'Updating precision and recall metrics...')
                                         # get the metrics and add them to df. save back the new file. then continue.
                                         # there is just one element in the df so we can do the operation below:
-                                        kde_eval_df['precision'] = precision
-                                        kde_eval_df['recall'] = recall
-                                        kde_eval_df['metrics'] = metrics
+                                        # kde_eval_df['precision'] = precision
+                                        # kde_eval_df['recall'] = recall
+                                        # kde_eval_df['metrics'] = metrics
+
+                                        # for 0.0
+                                        kde_eval_df['precision_0'] = precision_0
+                                        kde_eval_df['recall_0'] = recall_0
+                                        kde_eval_df['metrics_0'] = metrics_0
+
+                                        # for 0.025
+                                        kde_eval_df['precision_025'] = precision_025
+                                        kde_eval_df['recall_025'] = recall_025
+                                        kde_eval_df['metrics_025'] = metrics_025
+
+                                        # for 0.05
+                                        kde_eval_df['precision_05'] = precision_05
+                                        kde_eval_df['recall_05'] = recall_05
+                                        kde_eval_df['metrics_05'] = metrics_05
+
                                         kde_eval_df.to_csv(KDE_EVALUATION_DETAILS_FILENAME,index=False)
+                                        print(f'Saved update CSV file: {KDE_EVALUATION_DETAILS_FILENAME}')
+
                                         continue
 
 
@@ -158,7 +189,16 @@ if __name__=="__main__":
                                                 'kde':[KDE_MODEL_FILENAME],
                                                 'precision':[precision],
                                                 'recall': [recall],
-                                                'metrics' : [metrics]
+                                                'metrics' : [metrics],
+                                                'precision_0':[precision_0],
+                                                'recall_0': [recall_0],
+                                                'metrics_0' : [metrics_0],
+                                                'precision_05':[precision_05],
+                                                'recall_05': [recall_05],
+                                                'metrics_05' : [metrics_05],
+                                                'precision_025':[precision_025],
+                                                'recall_025': [recall_025],
+                                                'metrics_025' : [metrics_025]
                                                 })
 
                             df.to_csv(KDE_EVALUATION_DETAILS_FILENAME,index=False)
